# examples/Workshop LiveExample/Server/ASP.NET/src/IO.Swagger/Models/ArtefactDTO.cs
/*
 * Webtechnologien - Workshop
 *
 * Beispiel für den Workshop über Swagger und OpenAPI im Modul Webtechnologien
 *
 * OpenAPI spec version: 1.0.0
 * Contact: patrick.albus@smail.th-koeln.de
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{ 
    /// <summary>
    /// Artefact Data Transform Object
    /// </summary>
    [DataContract]
    public partial class ArtefactDTO : IEquatable<ArtefactDTO>
    { 
        /// <summary>
        /// Unique ID of Artefact
        /// </summary>
        /// <value>Unique ID of Artefact</value>
        [Required]
        [DataMember(Name="objectId")]
        public string ObjectId { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class ArtefactDTO {\n");
            sb.Append("  ObjectId: ").Append(ObjectId).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((ArtefactDTO)obj);
        }

        /// <summary>
        /// Returns true if ArtefactDTO instances are equal
        /// </summary>
        /// <param name="other">Instance of ArtefactDTO to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(ArtefactDTO other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    ObjectId == other.ObjectId ||
                    ObjectId != null &&
                    ObjectId.Equals(other.ObjectId)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (ObjectId != null)
                    hashCode = hashCode * 59 + ObjectId.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(ArtefactDTO left, ArtefactDTO right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(ArtefactDTO left, ArtefactDTO right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}