# examples/Workshop LiveExample/Server/NodeJS/api/openapi.yaml
openapi: 3.0.1
info:
  title: Webtechnologien - Workshop
  description: Beispiel für den Workshop über Swagger und OpenAPI im Modul Webtechnologien
  termsOfService: https://github.com/Narua2010
  contact:
    name: Patrick Albus
    url: https://github.com/Narua2010
    email: patrick.albus@smail.th-koeln.de
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: https://localhost:80/api
  description: Dev Server mit Port 80
- url: "https://localhost:8080/{version}/api"
  description: Dev Server mit Port 8080 und Versions Variable
  variables:
    version:
      default: v1
      enum:
      - rc
      - v1
paths:
  /artefacts:
    get:
      tags:
      - Artefacts
      summary: Example for GET request without parameters
      description: This API returns a list of artefacts.
      externalDocs:
        description: Externe Dokumentation
        url: https://th-koeln.github.io/mi-master-wtw/web-technologien/
      operationId: GetArtefacts
      responses:
        "200":
          description: List of artefacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArtefactDTO'
                x-content-type: application/json
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      x-swagger-router-controller: Artefacts
    post:
      tags:
      - Artefacts
      summary: Example for POST request
      operationId: AddArtefact
      requestBody:
        description: Data of new artefact
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtefactDTO'
        required: true
      responses:
        "201":
          description: Added artefact
          headers:
            Locatoin:
              description: URL where to get details of the created artefact
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtefactDTO'
          links:
            artefact:
              operationId: GetArtefactByObjectId
              parameters:
                objectId: response.body#/objectId
        "400":
          description: Invalid artefact data in request body
          content:
            text/html:
              schema:
                type: string
                description: Cause of the error
                example: Mandatory member 'location' is missing.
                x-content-type: text/html
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      x-swagger-router-controller: Artefacts
  /artefacts/{objectId}:
    get:
      tags:
      - Artefacts
      summary: Example for GET request with path paramenter.
      description: This API returns an artefact.
      externalDocs:
        description: Externe Dokumentation
        url: https://th-koeln.github.io/mi-master-wtw/web-technologien/
      operationId: GetArtefactByObjectId
      parameters:
      - name: objectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Artefact
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArtefactDTO'
                x-content-type: application/json
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      x-swagger-router-controller: Artefacts
components:
  schemas:
    ArtefactDTO:
      required:
      - objectId
      type: object
      properties:
        objectId:
          type: string
          description: Unique ID of Artefact
          nullable: false
      description: Artefact Data Transform Object
      example:
        objectId: objectId
    errorResponse:
      required:
      - error
      type: object
      properties:
        error:
          type: string
          description: Description of the error
  responses:
    errorResponse:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/errorResponse'